Fix issue with footnote number size.
The myriad package (created with otfinst) redefines the
\@makefnmark command to use \sustyle-sized numbers for the
footnotes, but doesn't actually put them into superscript, so they
end up on the baseline.

Loading myriad first, and adding the footnotefigures option to
MinionPro, allows MinionPro to define the appearance of footnote
numerals, overriding the broken otfinst behavior.  (Reported as a
bug to John Owens on 2011-02-01.)


git-svn-id: https://svn.math.hmc.edu/latex/codee/classes/trunk@364 386e5e47-19f0-0310-9baa-f3d740de32d5