Add support for lining figures for Libertine roman and Biolinum sans-serif fonts.
We now provide the "standard" \lnstyle and \textln commands (usually provided by nfssext-cfr).  Because of the way that the libertine package defines font familes, those commands use the Libertine roman fonts, and are thus not orthogonal the way that they would be with other fonts.  So to handle the Biolinum fonts we use in the headers, we've defined separate \sflnstyle and \textsfln commands that will use Biolinum's lining figures font family.

Of course should we change the typefaces in the future, we will have to ensure that we remove those definitions, and replace their use in various places with the standard commands.  (Note that for places where we use the sf variants, we would also need to add \sffamily or \textsf, as appropriate.